% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calitems.R
\name{calitems}
\alias{calitems}
\title{Allocate calibration items based on examinee's ability}
\usage{
calitems(yyy, abil)
}
\arguments{
\item{yyy}{a \code{\link{optical}} object; the output of a call \code{\link[=optical]{optical()}}}

\item{abil}{ability value of the examinee (relative to a population
with abilities following a standard normal distribution)}
}
\value{
A vector containing the item numbers allocated to the examinee with
the given ability 'abil'.
}
\description{
This function determine the calibration items to be allocated to an examinee with given
ability
}
\examples{
# Example No.1
# 2PL-models for two items; parameters (a, b)=(1.6, -1) and (1.6, 1), respectively
ip <- cbind(c(1.6, 1.6),c(-1, 1))
yyy <- optical(ip)
calitems(yyy, 0.27)

\dontrun{
# Example No.2
# 2PL-models for six items; parameters a=c(1.62, 0.66, 0.92, 0.90, 0.98, 1.40)
# and b=c(-0.47, -0.15, -1.71, 1.60, 0.62, -1.71), respectively.
# The calibration of these 6 items with 2PL-models is conducted in two blocks.

a<-c(1.62, 0.66,0.92,0.90,0.98,1.40)
b<-c(-0.47,-0.15,-1.71,1.60,0.62,-1.71)
ip<-cbind(a,b)
bid<-c(1,2,2,2,1,1)
yyy<-optical(ip,bid=bid)
calitems(yyy, 0.27)

}
}
\seealso{
\code{\link{drawdesign}} , \code{\link{drawdesign_allitems}}
}
