% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawdesign_allitems.R
\name{drawdesign_allitems}
\alias{drawdesign_allitems}
\title{Optimal calibration design for all items}
\usage{
drawdesign_allitems(yyy, linew = NULL, ablim = 4, colvec = NULL)
}
\arguments{
\item{yyy}{a \code{\link{optical}} object; the output of a call \code{\link[=optical]{optical()}}}

\item{linew}{line width for design}

\item{ablim}{ability limits; plots will be made in the range [-ablim, ablim]}

\item{colvec}{vector of color sequence for items (default is the R-default black, red, green, etc.)}
}
\value{
An optimal design plot is displayed for all items.
}
\description{
Create a plot with optimal design for each item. Optimal design is represented
by different colored lines for each item.
}
\examples{

# Example No.1
# 2PL-model for three items with parameters (a, b) equal to (1.6, -2), (1.6, 0.5),
# and (1.6, 2) for the first, second, and third items, respectively.

ip <- cbind(c(1.6, 1.6, 1.6), c(-2, 0.5, 2))
yyy <- optical(ip)
drawdesign_allitems(yyy)


\dontrun{
# Example No.2
# 2PL-models for six items; the parameters for these items are a=(1.62, 1.4, 0.98, 0.66, 0.92, 0.9),
# and b=(-0.47, -1.71, 0.62, -0.15, -1.71, 1.6), respectively.

ip <- cbind(c(1.62, 1.4, 0.98, 0.66, 0.92, 0.9), c(-0.47, -1.71, 0.62, -0.15, -1.71, 1.6))
bid <- c(1, 1, 1, 2, 2, 2)
yyy <- optical(ip, bid = bid, show_progress = 2)
drawdesign_allitems(yyy)

# Items of the same color that are in the same block.
drawdesign_allitems(yyy, colvec = bid)
}
}
