% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_opt.R
\name{linear_opt}
\alias{linear_opt}
\title{Linear Regression Optimisation}
\usage{
linear_opt(p_area)
}
\arguments{
\item{p_area}{a list of calibration data}
}
\value{
a list of three elements
}
\description{
Linear Regression Optimisation
}
\keyword{internal}
