% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise.R
\name{optimise}
\alias{optimise}
\title{Optimise Calibration Series}
\usage{
optimise(calibration_data, p = 0.6, quad = TRUE)
}
\arguments{
\item{calibration_data}{a \code{data.frame} or \code{tibble} of calibration data. The data should have two columns named, \code{area} and \code{concentration}.}

\item{p}{a numeric value (between 0 and 1) of the minimum percentage of calibration observations to utilise}

\item{quad}{logical; if \code{TRUE} then quadratic regression is also included alongside linear regression duringf the optimisation}
}
\value{
a list of three elements.
\itemize{
\item \code{type}: a character string of the regression type (linear or quadratic)
\item \code{score}: numeric value of the \code{OptiCal} score
\item \code{data}: a \code{data.frame} of the optimal set of calibration observations
}
}
\description{
Optimise Calibration Series
}
