% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.optical.R
\name{print.optical}
\alias{print.optical}
\title{print a optical object}
\usage{
\method{print}{optical}(x,digits=max(3, getOption("digits")-3),\dots)
}
\arguments{
\item{x}{optical object.}

\item{...}{other print arguments.}
}
\value{
No return value, called for side effects.
}
\description{
This function print the table of interval boundaries for
optimal design with items and probabilities.
}
\seealso{
\code{\link{optical}}
}
