% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quad_opt.R
\name{quad_opt}
\alias{quad_opt}
\title{Quadratic Regression Optimisation}
\usage{
quad_opt(p_area)
}
\arguments{
\item{p_area}{a list of calibration data}
}
\value{
a list of three elements
}
\description{
Quadratic Regression Optimisation
}
\keyword{internal}
