% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_block.R
\name{efficiency}
\alias{efficiency}
\title{Efficiency of optimal design}
\usage{
efficiency(
  yyy,
  uncert = FALSE,
  ipop,
  oc = "D",
  L = NULL,
  items = FALSE,
  integ = TRUE
)
}
\arguments{
\item{yyy}{a \code{\link{optical}} object; the output of a call \code{\link[=optical]{optical()}}}

\item{uncert}{if false (default), abilities are assumed to be known; if true,
handling of uncertainties of Bjermo et al. (2021) is used.}

\item{ipop}{matrix with item parameters for operational items
(used if uncert=TRUE, only).}

\item{oc}{optimality criterion: "D" (D-optimality, default),
"I" (I-optimality with standard normal weight function), "A" (A-optimality).}

\item{L}{L-matrix (not used for D-optimality)}

\item{items}{If false (default), only total block efficiency is returned. If true,
for each block, criteria for optimal and random, and the efficiency
for each item are reported in each column of output, except for 1-pl
model where each column represents the parameter efficiency. The last column
shows total criteria and efficiency. D-, L-, I-, A-optimality.}

\item{integ}{if true (default), integrate() is used for computation of partial
information matrices; if false, Riemann rule is used.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  A numerical value is displayed.
}\if{html}{\out{</div>}}
}
\description{
This function computes the efficiency of the D, I, and A optimal designs
compared to the random design for each block.
}
\examples{

# Example No.1
# 2PL-models for two items; parameters (a, b)=(1.6, -1) and (1.6, 1), respectively

a<-c(1.6, 1.6); b<-c(-1, 1)
ip <- cbind(a,b)
yyy <- optical(ip)

# Efficiency of A-optimal design compared to random design
efficiency(yyy, oc="A")

\dontrun{

# Example No.2
# 2PL-models for six items; the parameters for these items are a=(1.62, 1.4,0.98,0.66,0.92,0.9),
# and b=(-0.47, -1.71, 0.62,-0.15,-1.71,1.6), respectively.

a <- c(1.62, 1.4, 0.98, 0.66, 0.92, 0.9)
b <- c(-0.47, -1.71, 0.62, -0.15, -1.71, 1.6)
ip <- cbind(a, b)
bid <- c(1, 1, 1, 2, 2, 2)
yyy <- optical(ip, BID = bid, show_progress = 2)

# Efficiency of D-optimal design compared to random design
efficiency(yyy, oc = "D")

# Efficiency of A-optimal design compared to random design
efficiency(yyy, oc = "A")

# Efficiency of I-optimal design compared to random design
efficiency(yyy, oc = "I")
}
}
\seealso{
\code{\link{optical}}
}
