% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{efficiency}
\alias{efficiency}
\title{Efficiency of optimal design}
\usage{
efficiency(
  t,
  ip,
  uncert = FALSE,
  ipop,
  xi,
  oc,
  L = NULL,
  items = FALSE,
  integ = TRUE
)
}
\arguments{
\item{t}{vector of abilities}

\item{ip}{matrix with item parameters for all items (number of rows determines
number of items number of column 2 (2PL) or 3 (3PL or mixed 2/3-PL with NA for
2PL-items in third column).}

\item{uncert}{if false (default), abilities are assumed to be known; if true,
handling of uncertainties of Bjermo et al. (2021) is used.}

\item{ipop}{matrix with item parameters for operational items
(used if uncert=TRUE, only).}

\item{xi}{vector of optimal design}

\item{oc}{optimality criterion: "D" (D-optimality, default),
"I" (I-optimality with standard normal weight function), "A" (A-optimality).}

\item{L}{L-optimality}

\item{items}{if true (default), criteria for optimal and random and the efficiency
for each item are reported in each column of output. Last column are then
total criteria and efficiency. D-, L-, I-, A-optimality}

\item{integ}{if true (default), integrate() is used for computation of partial
information matrices; if false, Riemann rule is used.}
}
\description{
Efficiency of optimal design
}
