% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optical.R
\name{optical}
\alias{optical}
\title{Optimal item calibration}
\usage{
optical(
  ip,
  oc = "D",
  uncert = FALSE,
  ipop,
  imf = c(0.005, 0.01, 0.02, 0.05, 0.1, 0.2, 0.45),
  maxiter = rep(300, 6),
  eps = rep(0.002, 6),
  nnn = c(0, 50, 50, 200, 200, 200),
  nsp = c(0.001, 1e-04, 1e-04, 1e-05, 1e-05, 1e-05),
  sss = 0.001,
  falpha = 1.08,
  sdr = TRUE,
  ig = 3,
  ex = 0,
  integ = TRUE
)
}
\arguments{
\item{ip}{matrix with item parameters for all items (number of rows determines
number of items number of column 2 (2PL) or 3 (3PL or mixed 2/3-PL with NA for
2PL-items in third column).}

\item{oc}{optimality criterion: "D" (D-optimality, default),
"I" (I-optimality with standard normal weight function), "A" (A-optimality).}

\item{uncert}{if false (default), abilities are assumed to be known; if true,
handling of uncertainties of Bjermo et al. (2021) is used.}

\item{ipop}{matrix with item parameters for operational items
(used if uncert=TRUE, only).}

\item{imf}{the vector of step-lengths; default
c(0.005, 0.01, 0.02, 0.05, 0.1, 0.2, 0.45).}

\item{maxiter}{maximal number of iterations in each inner loop, the length
of this vector defines the number of outer loops.}

\item{eps}{convergence criterion (maximum violation of eq.th.), vector with
value for each inner loop, but the same number for all inner loops is
recommended.}

\item{nnn}{number of new nodes added at each position in the adaptive grid,
vector with value for each inner loop (nnn [1] not used).}

\item{nsp}{node spacing between new nodes, vector with value for each inner
loop (nsp [1] is the spacing between nodes of the starting grid).}

\item{sss}{step size stopping criterion.}

\item{falpha}{factor alpha for adjusting the step size vector.}

\item{sdr}{stop if design repeated (flag TRUE/FALSE).}

\item{ig}{inner grid between -ig and ig.}

\item{ex}{intervals of size < ex will be removed (consolidate);
if ex=0, no consolidation will be done.}

\item{integ}{if true (default), integrate() is used for computation of partial
information matrices; if false, Riemann rule is used.}
}
\value{
Result of this function is a list with following instances:
\item{dd}{directional derivatives of optimal solution.}
\item{xi}{optimal solution.}
\item{t}{final grid of ability values which was used.}
\item{viomax}{largest violation of eq.th. from final solution (if < eps, alg.
has converged, otherwise not).}
\item{h1}{interval boundaries for optimal solution.}
\item{mooiter}{monitoring iterations; information about each iteration to produce
convergence plots.}
\item{time}{running time of algorithm in minutes.}
\item{oc}{optimality criterion ("D", "I", "A", "L").}
\item{L}{L-matrix (not for D-optimality).}
}
\description{
Calibrate the items follow the model 2PL, 3PL, or mixture of 2PL and 3PL
model.
}
\examples{
# 1PL-models with common discrimination parameter
ip <- cbind(c(1.6, NA), c(-1, 1))

yyy <- optical(ip, oc="D", uncert=FALSE, ipop,
               imf=c(0.005, 0.01, 0.02, 0.05, 0.1, 0.2, 0.45),
               maxiter=rep(300, 6), eps=rep(0.002, 6),
               nnn=c(0, 50, 50, 200, 200, 200),
               nsp=c(0.001, 0.0001, 0.0001, 0.00001, 0.00001, 0.00001),
               sss=0.001, falpha=1.08, ig=3, ex=0)

yyy$h1
}
\references{
Ul Hassan and Miller (2021). An exchange algorithm for optimal calibration
of items in computerized achievement tests.\emph{Computational Statistics and
Data Analysis, 157}: 107177. \doi{https://doi.org/10.1016/j.csda.2021.107177}

Bjermo, Fackle-Fornius, and Miller (2021). Optimizing Calibration Designs with
Uncertaintyin Abilities. Manuscript. \href{https://urn.kb.se/resolve?urn=urn\%3Anbn\%3Ase\%3Asu\%3Adiva-198065}{urn:nbn:se:su:diva-198065}
}
\seealso{
\code{\link{drawdesign}}, \code{\link{convergenceplot}}
}
\author{
Mahmood Ul Hassan (\email{scenic555@gmail.com});
Frank Miller (\email{frank.miller@stat.su.se})
}
